barbearia/usuarios/models.py
from django.db import models
from django.contrib.auth.models import User
from django.db.models.signals import post_save
from django.dispatch import receiver
from PIL import Image
import os
import sys

class Profile(models.Model):
    """Modelo para perfil de usuário com informações adicionais"""
    user = models.OneToOneField(User, on_delete=models.CASCADE)
    cpf = models.CharField(max_length=14, unique=True, verbose_name='CPF')
    telefone = models.CharField(max_length=15, verbose_name='Telefone')
    foto = models.ImageField(default='default.jpg', upload_to='profile_pics', verbose_name='Foto')
    data_cadastro = models.DateTimeField(auto_now_add=True, verbose_name='Data de Cadastro')
    
    def __str__(self):
        return f'Perfil de {self.user.username}'
    
    def save(self, *args, **kwargs):
        super().save(*args, **kwargs)
        
        # Pular o redimensionamento de imagem durante os testes
        if 'test' in sys.argv:
            return
            
        # Redimensionar imagem de perfil se o arquivo existir
        if os.path.exists(self.foto.path):
            try:
                img = Image.open(self.foto.path)
                if img.height > 300 or img.width > 300:
                    output_size = (300, 300)
                    img.thumbnail(output_size)
                    img.save(self.foto.path)
            except Exception:
                # Ignorar erros de processamento de imagem
                pass

# Criar perfil automaticamente quando um usuário for criado
@receiver(post_save, sender=User)
def create_profile(sender, instance, created, **kwargs):
    if created:
        Profile.objects.create(user=instance)

@receiver(post_save, sender=User)
def save_profile(sender, instance, **kwargs):
    instance.profile.save()
