requirements.txt
Django==5.2
django-crispy-forms==2.4
Pillow==11.1.0
